moved channelgraph to network
